Move all CfgACEArsenalStats entries to their own file
