+++ addons/hearing/CfgACEArsenalStats.hpp
@@ -3,7 +3,7 @@ class CfgACEArsenalStats {
     class hearingProtection: statBase {
         scope = 2;
         priority = 2;
-        stats[]= {"ACE_hearing_protection"};
+        stats[]= {QGVAR(protection)};
         displayName= CSTRING(statHearingProtection);
         showBar = 1;
         barStatement = QUOTE([ARR_3((_this select 0) select 0, _this select 1, [ARR_3([ARR_2(0, 1)], [ARR_2(0.01, 1)], false)])] call EFUNC(arsenal,statBarStatement_default));
@@ -12,7 +12,7 @@ class CfgACEArsenalStats {
     class volumeMuffling: statBase {
         scope = 2;
         priority = 1;
-        stats[]= {"ACE_hearing_lowerVolume"};
+        stats[]= {QGVAR(lowerVolume)};
         displayName= CSTRING(statHearingLowerVolume);
         showBar = 1;
         barStatement = QUOTE([ARR_3((_this select 0) select 0, _this select 1, [ARR_3([ARR_2(0, 1)], [ARR_2(0.01, 1)], false)])] call EFUNC(arsenal,statBarStatement_default))
