+++ addons/overpressure/CfgACEArsenalStats.hpp
@@ -3,7 +3,7 @@ class CfgACEArsenalStats {
     class backblastAngle: statBase {
         scope = 2;
         priority = 2;
-        stats[]= {"ace_overpressure_angle"};
+        stats[]= {QGVAR(angle)};
         displayName= CSTRING(statBackblastAngle);
         showText= 1;
         textStatement = QUOTE(params [ARR_2('_stat', '_config')]; format [ARR_2('%1°', getNumber (_config >> _stat select 0))]);
@@ -12,7 +12,7 @@ class CfgACEArsenalStats {
     class backblastRange: statBase {
         scope = 2;
         priority = 1;
-        stats[]= {"ace_overpressure_range"};
+        stats[]= {QGVAR(range)};
         displayName= CSTRING(statBackblastRange);
         showText= 1;
         textStatement = QUOTE(params [ARR_2('_stat', '_config')]; private _blastRangeStat = getNumber (_config >> _stat select 0); format [ARR_3('%1m (%2ft)', _blastRangeStat, (_blastRangeStat / 0.3048) toFixed 1)]);
